requirements.txt
flask==2.0.1
google-api-python-client==2.47.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
Werkzeug==2.0.1
gunicorn==20.1.0
python-dotenv==0.19.0
requests==2.31.0
Pillow==10.0.0 